Store only Order/range/counterCurrency
