+++ apps/ledger-live-desktop/src/renderer/screens/market/hooks/useMarketCoin.ts
@@ -9,14 +9,11 @@ import {
 import { Page, useMarketActions } from "./useMarketActions";
 import { useCallback } from "react";
 import { useParams } from "react-router";
-import {
-  addStarredMarketCoins,
-  removeStarredMarketCoins,
-  setMarketOptions,
-} from "~/renderer/actions/market";
-import { marketParamsSelector, starredMarketCoinsSelector } from "~/renderer/reducers/market";
+import { setMarketOptions } from "~/renderer/actions/market";
+import { marketParamsSelector } from "~/renderer/reducers/market";
 import { useFetchCurrencyAll } from "@ledgerhq/live-common/exchange/swap/hooks/index";
-import { localeSelector } from "~/renderer/reducers/settings";
+import { localeSelector, starredMarketCoinsSelector } from "~/renderer/reducers/settings";
+import { removeStarredMarketCoins, addStarredMarketCoins } from "~/renderer/actions/settings";
 
 export const useMarketCoin = () => {
   const marketParams = useSelector(marketParamsSelector);
